pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anyio==4.9.0",
    "appnope==0.1.4",
    "asttokens==3.0.0",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.1",
    "comm==0.2.2",
    "debugpy==1.8.14",
    "decorator==5.2.1",
    "executing==2.2.0",
    "greenlet==3.2.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "ipykernel==6.29.5",
    "ipython==9.2.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "matplotlib>=3.10.1",
    "matplotlib-inline==0.1.7",
    "nest-asyncio==1.6.0",
    "numpy==2.2.5",
    "packaging==25.0",
    "pandas==2.2.3",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pillow==11.2.1",
    "platformdirs==4.3.7",
    "playwright==1.51.0",
    "prompt-toolkit==3.0.51",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyarrow==20.0.0",
    "pyee==12.1.1",
    "pygments==2.19.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyzmq==26.4.0",
    "requests==2.32.3",
    "six==1.17.0",
    "sniffio==1.3.1",
    "stack-data==0.6.3",
    "tiktokapi==7.1.0",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typing-extensions==4.13.2",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "wcwidth==0.2.13",
]
